Rename to used class
